add component tests to fluxsvc
The goal is to fully test the fluxsvc API. (Thereby decoupling fluxctl and fluxd from fluxsvc).
The fluxsvc server will be started and all dependencies are in memory.
Reduced key gen size to 2048. Taking too long for 4096. Users can regen their keys periodically if they are worried.
Refactored so that transport invoke methods are exported, for reuse in tests.