Give objects a source path
This adds a `Source()` method object so it can be output in diffs, and
used as a path for loading the full file otherwise.

One subtle change is that a blank namespace is no longer defaulted to
"default" when unmarshaling; it's left as-is, and defaulted when
constructing the ID. This doesn't really change anything for Nodes and
Namespaces, which should both have empty namespace entries.